requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
fpdf==1.7.2
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
newsapi-python==0.2.7
oauthlib==3.2.2
Pillow==9.4.0
PyPDF2==3.0.1
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
python-telegram-bot==20.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.9
telebot==0.0.5
tweepy==4.12.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
